+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # kgaut_tools 2.0.x-dev
 - Add new submodule kgaut_tools_formatters
 - kgaut_tools_formatters - Add StringTitleFormatter
+- kgaut_tools_formatters - Add LinkButtonFormatter
 
 # kgaut_tools 2.0.5 27/01/2021
 - EntityStatusTrait - fix method isPublished()